+++ src/templates/rude/ajax/rude-ajax-user.php
@@ -6,20 +6,8 @@ class ajax_user
 {
 	public static function has_access()
 	{
-	    $role_name = get(RUDE_FIELD_ROLE, $_SESSION);
-
-		if ($role_name === false)
-		{
-			return false;
-		}
-
-		$role = roles::get_by_name($role_name);
-
-		if ($role === null)
-		{
-			return false;
-		}
-		return $role->allow_user_management === '1';
+		$allow_user_management = get(RUDE_FIELD_ALLOW_USER_MANAGEMENT, $_SESSION);
+		return $allow_user_management === '1';
 	}
 
     public static function add()
@@ -90,6 +78,11 @@ class ajax_user
 
 		$username = get(RUDE_FIELD_USERNAME);
 
+		if ($username === get(RUDE_FIELD_USERNAME, $_SESSION))
+		{
+			die();
+		}
+
 		users::delete($username);
 	}
 
@@ -109,6 +102,11 @@ class ajax_user
 
 	public static function html_form_add()
 	{
+		if (!ajax_user::has_access())
+		{
+			die();
+		}
+
 		?>
 		<html>
 		<? require_once(RUDE_TEMPLATE_DIR . '/rude-header.php') ?>
@@ -216,6 +214,11 @@ class ajax_user
 
 	public static function html_form_edit()
 	{
+		if (!ajax_user::has_access())
+		{
+			die();
+		}
+
 		?>
 		<html>
 		<? require_once(RUDE_TEMPLATE_DIR . '/rude-header.php') ?>
@@ -331,6 +334,18 @@ class ajax_user
 
 	public static function html_form_delete()
 	{
+		if (!ajax_user::has_access())
+		{
+			die();
+		}
+
+		$username = get(RUDE_FIELD_USERNAME);
+
+		if ($username === get(RUDE_FIELD_USERNAME, $_SESSION))
+		{
+			die();
+		}
+
 		?>
 		<html>
 		<? require_once(RUDE_TEMPLATE_DIR . '/rude-header.php') ?>
@@ -340,10 +355,6 @@ class ajax_user
 			<form id="form" name="form" method="post" action="index.php?task=<?= RUDE_TASK_USER_DELETE ?>">
 				<h1>Удаление пользователя</h1>
 				<p class="red">Внимание! Данная операция необратима!</p>
-
-				<?
-				$username = get(RUDE_FIELD_USERNAME);
-				?>
 
 				Вы точно уверены, что хотите удалить пользователя <b>"<?= $username ?></b>"?
 
@@ -402,7 +413,9 @@ class ajax_user
 					<td><?= $user->role ?></td>
 					<td>
 						<a href="<?= url::ajax(RUDE_TASK_AJAX_USER_EDIT_FORM) . url::param(RUDE_FIELD_USERNAME, $user->username) ?>" class="fancybox-users"><img src="src/icons/edit.png" class="padding-small" title="<?= RUDE_TEXT_EDIT ?>" /></a>
+					<? if ($user->username !== get(RUDE_FIELD_USERNAME, $_SESSION)) : ?>
 						<a href="<?= url::ajax(RUDE_TASK_AJAX_USER_DELETE_FORM) . url::param(RUDE_FIELD_USERNAME, $user->username) ?>" class="fancybox-users-small"><img src="src/icons/remove.png" class="padding-small" title="<?= RUDE_TEXT_DELETE_SELECTED ?>" /></a>
+					<? endif; ?>
 					</td>
 				</tr>
 			<?
